Refactor pi library to use a configurable object instead of package globals
 - fold bpp_spigot.go content into pi.go
 - remove the single digit function;
